fix conat unit tests after changing definition of sticky
