+++ src/packages/backend/conat/test/server/pubsub.test.ts
@@ -29,7 +29,7 @@ import {
   runValkey,
 } from "@cocalc/backend/conat/test/setup";
 import { STICKY_QUEUE_GROUP } from "@cocalc/conat/core/client";
-import { waitForSubscription } from "./valkey.test";
+import { waitForSubscription } from "./util";
 
 // should be several thousand, so 250 seems reasonable as a cutoff to indicate
 // things are horribly wrong
