+++ src/packages/backend/conat/test/sync/dstream.test.ts
@@ -14,6 +14,8 @@ import { connect, before, after, wait } from "@cocalc/backend/conat/test/setup";
 
 beforeAll(before);
 
+jest.setTimeout(10000);
+
 describe("create a dstream and do some basic operations", () => {
   let s;
